Added decimal, updated float, comment operators, better boundaries
Abstracted pattern matching for boolean context and numeric value
contexts  (number, decimal, floating-point-number) from <match>
tags to a <define-regex> to allow for greater clarity.

Previously some matches would still highlight if running on from
invalid prefix chars, eg %pi might not be highlighted in the string
w%pi but it might be highlighted in #%pi. Root of the problem appears
to be Scilab's usage of some characters. %, ? and # are all valid
"word chars" and therefore these have been added to a regex roughly
analagous to the \w special char, and the regex employed as a
negative lookbehind/ahead to ensure boundaries are adhered to. It
seems like \%[ and \%] are designed for this, but despite giving them
a custom definition in <keyword-char-class> (to differentiate from
their default def. as \b) they don't appear to have any effect in
regex patterns, perhaps only keywords?

The improved boundaries have been implemented for boolean, number,
decimal and float contexts.